restore stations from url
